Refactor data table
Data table with custom styling now its own component.
